backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
groq
